+++ crates/cli-tools/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0-git
 
-<!-- Increment to skip CHANGELOG.md test: 15 -->
+<!-- Increment to skip CHANGELOG.md test: 16 -->